+++ src/kll/kll.cpp
@@ -85,7 +85,7 @@ namespace qsbd {
 	double kll<T>::cdf(T elem){
 		if (not this->N) return 0.0;
 
-		return this->query(elem) / this->N;
+		return this->query(elem) / (double) this->N;
 	} 
 
 	template<class T>
